% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-action.R
\name{act_series}
\alias{act_series}
\title{Sequences of actions}
\usage{
act_series(..., repeating = 0)
}
\arguments{
\item{...}{One or more \code{\link{action}} objects to execute in sequence.}

\item{repeating}{Number of times to repeat the entire series
(0 = run once, 1 = run twice, etc.).}
}
\value{
An object of class \code{act_series}, which itself is an \code{\link{action}}.
}
\description{
\code{act_series()} combines multiple \code{\link{action}} objects into a
single composite action that runs them in sequence, optionally
repeating the entire series.
}
\details{
An \code{act_series} executes each action in sequence, advancing to the
next only after the current action completes. The total duration
is the sum of all individual action durations.
}
\examples{
# Sequence: move for 1 second, then scale for 0.5 seconds
series <- act_series(
  translate(to = pos(5, 5), time = time(1)),
  scale2(size = 2, time = time(0.5))
)

}
\seealso{
\code{\link{action}}, \code{\link{time}}, \code{\link[=translate]{translate()}}, \code{\link[=scale2]{scale2()}}
}
