% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-action.R
\name{action}
\alias{action}
\title{Action objects}
\usage{
action(func, time)
}
\arguments{
\item{func}{A function with signature \verb{function(obj, time)} that
performs the animation. The \code{obj} parameter is the shape being
animated, and \code{time} is a \code{\link{time}} object tracking progress.}

\item{time}{A \code{\link{time}} object controlling duration, easing, and
repetition of the action.}
}
\value{
An object of class \code{action}.
}
\description{
\code{action()} wraps a simple \verb{(obj, time)} function into an object that
can be scheduled and combined with other actions. Actions are
associated with \code{\link[=time]{time()}} objects and are attached to \link{shape} instances
via the \verb{@action} property.
}
\section{Methods}{

\itemize{
\item \verb{$then(next_action)}: Chain another action to run after this one.
Returns an \code{\link{act_series}}.
\item \verb{$finally(final_action)}: Alias for \verb{$then()}.
\item \verb{$reset()}: Reset the time state to the beginning.
\item \verb{$repeating(times)}: Set the number of repetitions.
\item \verb{$is_done}: Logical; whether the action is complete.
\item \verb{$left}: Remaining time in the current cycle.
}
}

\seealso{
\code{\link[=time]{time()}}, \link{shape}, \link{act_series}, \code{\link[=into_action]{into_action()}}
}
