% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render-helpers.R
\name{center}
\alias{center}
\title{Get the center of a shape}
\usage{
center(obj)
}
\arguments{
\item{obj}{A \code{\link{shape}} object.}
}
\value{
A \code{\link{pos}} object representing the center.
}
\description{
\code{center()} computes the center point of a \code{\link{shape}}'s bounding box.
}
\seealso{
\code{\link[=get_extent]{get_extent()}}, \code{\link[=lc]{lc()}}, \code{\link[=rc]{rc()}}, \code{\link[=tc]{tc()}}, \code{\link[=bc]{bc()}}
}
\keyword{internal}
