% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-ease.R
\name{check_valid_ease}
\alias{check_valid_ease}
\title{Validate and check easing functions}
\usage{
check_valid_ease(f, name)
}
\arguments{
\item{f}{A function to validate.}

\item{name}{Name of the function for error messages.}
}
\value{
\code{TRUE} invisibly if valid; stops with error otherwise.
}
\description{
Validate and check easing functions
}
\keyword{internal}
