% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-actions.R
\name{color}
\alias{color}
\title{Color transition action}
\usage{
color(color_from, color_to, time, ...)
}
\arguments{
\item{color_from}{Starting color (any format accepted by \code{\link{class_color}}).}

\item{color_to}{Ending color.}

\item{time}{A \code{\link{time}} object controlling duration and easing.}
}
\value{
An \code{\link{action}} object.
}
\description{
\code{color()} creates an action that smoothly transitions a \code{\link{shape}}'s
color from one to another over time, using linear interpolation.
}
\examples{
w <- window(bl = 0, tr = pos(10, 10))
rect_obj <- rect(
  width = 2, height = 2, color = "red",
  trans = transform(pos(5, 5))
)
w@child(rect_obj)
rect_obj@action(color("red", "blue", time = time(1)))

}
\seealso{
\code{\link[=colors]{colors()}}, \code{\link{action}}, \code{\link{time}}, \code{\link{class_color}}
}
