% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-actions.R
\name{despawn}
\alias{despawn}
\title{Remove a shape from its parent}
\usage{
despawn(time = 0)
}
\arguments{
\item{time}{A \code{\link{time}} object controlling when the removal occurs.
Default 0 means remove immediately.}
}
\value{
An \code{\link{action}} object.
}
\description{
\code{despawn()} creates an action that detaches a shape from its
parent when the action completes.
}
\seealso{
\code{\link[=spawn]{spawn()}}, \code{\link[=spawn_sibling]{spawn_sibling()}}, \code{\link[=spawn_from]{spawn_from()}}
}
\keyword{internal}
