% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-ease.R
\name{elinear}
\alias{elinear}
\alias{ecos}
\alias{esin}
\alias{equad}
\alias{ecubic}
\title{Ease functions}
\usage{
elinear(x)

ecos(t)

esin(t)

equad(t)

ecubic(t)
}
\arguments{
\item{t}{A numeric value in the closed interval \[0, 1\] representing
normalized time.}
}
\value{
Numeric value in \[0, 1\] when \code{t} is in \[0, 1\].
}
\description{
Pre-defined easing functions for animation timing. These can be used
directly or combined with \code{\link[=ease]{ease()}} to create custom easing curves.
}
\details{
Standard easing functions:
\itemize{
\item \code{elinear}: Linear easing (identity function)
\item \code{ecos}: Cosine-based ease-in curve
\item \code{esin}: Sine-based ease-in curve
\item \code{equad}: Quadratic ease-in curve (t^2)
\item \code{ecubic}: Cubic ease-in curve (t^3)
}

All easing functions take a numeric value \code{t} in the closed interval
\[0, 1\] and return a value in \[0, 1\], mapping start to end of animation.
Values of \code{t} outside \[0, 1\] are allowed but may produce results
outside \[0, 1\] depending on the specific easing function.
}
\examples{
# Linear easing
elinear(0.5)

# Cubic easing
ecubic(0.5)

# Plot easing curve
t <- seq(0, 1, length.out = 100)
plot(t, Vectorize(ecubic)(t), type = "l")

}
