% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-action.R
\name{enforce_zero}
\alias{enforce_zero}
\title{Enforce zero on near-zero values}
\usage{
enforce_zero(old_dt, new_dt)
}
\arguments{
\item{old_dt}{Original delta time.}

\item{new_dt}{New delta time.}
}
\value{
Numeric value, coerced to exactly 0 if nearly zero.
}
\description{
Enforce zero on near-zero values
}
\keyword{internal}
