% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-action.R
\name{ensub}
\alias{ensub}
\title{Substitute expressions in an environment}
\usage{
ensub(expr, env)
}
\arguments{
\item{expr}{An expression to substitute.}

\item{env}{An environment in which to perform substitution.}
}
\value{
The substituted expression.
}
\description{
Substitute expressions in an environment
}
\keyword{internal}
