% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-action.R
\name{expr_map}
\alias{expr_map}
\title{Expression mapping for action constructors}
\usage{
expr_map(...)
}
\arguments{
\item{...}{Named arguments mapping parameter names to expressions.}
}
\value{
An object of class \code{expr_map}.
}
\description{
\code{expr_map()} is a small helper class used by \code{\link[=into_action]{into_action()}} to
describe how arguments in a user-facing action constructor
are rewritten into calls to a lower-level implementation.
}
\details{
Most users should not need to work with \code{expr_map} directly.
}
\keyword{internal}
