% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render-helpers.R
\name{get_extent}
\alias{get_extent}
\title{Get extent (bounding box) of a shape}
\usage{
get_extent(obj)
}
\arguments{
\item{obj}{A \code{\link{shape}} object.}
}
\value{
A \code{\link{pos}} object with two elements: \verb{[1]} is the bottom-left
corner (min x, min y), and \verb{[2]} is the top-right corner (max x, max y).
}
\description{
\code{get_extent()} computes the bounding box of a \code{\link{shape}} by finding
the minimum and maximum x and y coordinates of all positions in
the shape hierarchy.
}
\seealso{
\code{\link[=get_positions]{get_positions()}}, \code{\link[=center]{center()}}, \code{\link[=lc]{lc()}}, \code{\link[=rc]{rc()}}, \code{\link[=tc]{tc()}}, \code{\link[=bc]{bc()}}
}
\keyword{internal}
