% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-action.R
\name{get_methods}
\alias{get_methods}
\title{Get methods for a generic}
\usage{
get_methods(generic, name = deparse(substitute(generic)))
}
\arguments{
\item{generic}{A generic function or name.}

\item{name}{String name of the generic.}
}
\value{
A list of method implementations.
}
\description{
Get methods for a generic
}
\keyword{internal}
