% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render-helpers.R
\name{get_positions}
\alias{get_positions}
\title{Get positions of a shape or object}
\usage{
get_positions(obj)
}
\arguments{
\item{obj}{A \code{\link{shape}} or other object containing positions.}
}
\value{
A \code{\link{pos}} vector containing all positions.
}
\description{
\code{get_positions()} recursively collects all \code{\link{pos}} objects from a
shape and its children. Used internally to compute bounding boxes
and extents.
}
\keyword{internal}
