% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-action.R
\name{into_action}
\alias{into_action}
\title{Convert a function into an action}
\usage{
into_action(func, remap, obj_arg = names(formals(func))[1])
}
\arguments{
\item{func}{A function to wrap. It should normally take a shape-like
object as its first argument.}

\item{remap}{An \code{\link{expr_map}} object describing how arguments
in the returned action constructor are translated into
arguments of \code{func}. Captured arguments are forced before
the action is created to ensure proper scoping.}

\item{obj_arg}{Name of the argument in \code{func} that receives
the animated object. Defaults to the name of the first
formal argument of \code{func}.}
}
\value{
A new function that, when called with \verb{time=} and any
additional arguments, returns an \code{\link{action}} object.
}
\description{
Convert a function into an action
}
\examples{
# This is how translate() is implemented:
# translate <- into_action(obj_translate,
#   expr_map(to = to, local = FALSE))

}
\seealso{
\code{\link{action}}, \code{\link{expr_map}}
}
