% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-pos.R
\name{is_pos}
\alias{is_pos}
\title{Check if an object is a position vector}
\usage{
is_pos(x)
}
\arguments{
\item{x}{Object to check.}
}
\value{
Logical scalar.
}
\description{
Check if an object is a position vector
}
