% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-pos.R
\name{is_scalar_pos}
\alias{is_scalar_pos}
\title{Check if an object is a scalar position}
\usage{
is_scalar_pos(x)
}
\arguments{
\item{x}{Object to check.}
}
\value{
Logical scalar.
}
\description{
\code{is_scalar_pos()} checks if \code{x} is a position with exactly
one element.
}
