% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-ease.R
\name{is_valid_ease}
\alias{is_valid_ease}
\title{Check if a function is a valid ease function}
\usage{
is_valid_ease(f)
}
\arguments{
\item{f}{A function to validate.}
}
\value{
Logical scalar.
}
\description{
Check if a function is a valid ease function
}
\keyword{internal}
