% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render-helpers.R
\name{lc}
\alias{lc}
\alias{rc}
\alias{tc}
\alias{bc}
\title{Get edge centers of a shape}
\usage{
lc(obj)

rc(obj)

tc(obj)

bc(obj)
}
\arguments{
\item{obj}{A \code{\link{shape}} object.}
}
\value{
A \code{\link{pos}} object representing the requested edge center.
}
\description{
These functions return the center points of each edge of a shape's
bounding box.
}
\details{
\itemize{
\item \code{lc()}: left center
\item \code{rc()}: right center
\item \code{tc()}: top center
\item \code{bc()}: bottom center
}
}
\seealso{
\code{\link[=tl]{tl()}}, \code{\link[=tr]{tr()}}, \code{\link[=bl]{bl()}}, \code{\link[=br]{br()}}, \code{\link[=center]{center()}}
}
\keyword{internal}
