% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-action.R
\name{match_call}
\alias{match_call}
\title{Match function call}
\usage{
match_call(obj)
}
\arguments{
\item{obj}{Object for which to match the call.}
}
\value{
A call object.
}
\description{
Match function call
}
\keyword{internal}
