% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-transform.R
\name{obj_anchor}
\alias{obj_anchor}
\title{Query anchor of an object}
\usage{
obj_anchor(obj)
}
\arguments{
\item{obj}{Object to query.}
}
\value{
A \code{\link{pos}} object.
}
\description{
\code{obj_anchor()} retrieves the anchor point used for transformations.
}
