% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-transform.R
\name{obj_angle}
\alias{obj_angle}
\title{Query rotation angle of an object}
\usage{
obj_angle(obj)
}
\arguments{
\item{obj}{Object to query.}
}
\value{
Numeric angle in degrees.
}
\description{
\code{obj_angle()} retrieves the rotation angle of a shape or transform.
}
