% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{obj_color_between}
\alias{obj_color_between}
\title{Interpolate color between two values}
\usage{
obj_color_between(obj, color_from, color_to, amount)
}
\arguments{
\item{obj}{A \code{\link{shape}} object to recolor.}

\item{color_from}{Starting color.}

\item{color_to}{Ending color.}

\item{amount}{Numeric in \[0, 1\] giving interpolation fraction
(0 = fully \code{color_from}, 1 = fully \code{color_to}).}
}
\value{
The modified \code{shape}, invisibly.
}
\description{
\code{obj_color_between()} sets a shape's color to an interpolated value
between two colors. This is used internally by the \code{\link[=color]{color()}} action.
}
\seealso{
\code{\link[=color]{color()}}, \link{class_color}
}
\keyword{internal}
