% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-actions.R
\name{obj_offset}
\alias{obj_offset}
\title{Offset a shape's position}
\usage{
obj_offset(obj, by)
}
\arguments{
\item{obj}{A \code{\link{shape}} object.}

\item{by}{A \code{\link{pos}} offset to apply.}
}
\value{
The modified \code{shape}, invisibly.
}
\description{
\code{obj_offset()} moves a shape by a given delta.
}
\keyword{internal}
