% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-transform.R
\name{obj_pos}
\alias{obj_pos}
\title{Query position of an object}
\usage{
obj_pos(obj, local = FALSE)
}
\arguments{
\item{obj}{Object to query.}

\item{local}{If \code{FALSE} (default), return global (world) coordinates;
if \code{TRUE}, return local offset relative to the anchor.}
}
\value{
A \code{\link{pos}} object.
}
\description{
\code{obj_pos()} retrieves the position of a \code{\link{shape}} or \code{\link{transform}}.
}
