% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate.R
\name{obj_rotate}
\alias{obj_rotate}
\title{Rotate a shape or transform}
\usage{
obj_rotate(
  obj,
  around,
  ...,
  radians = degrees * pi/180,
  degrees = radians * 180/pi,
  local = FALSE
)
}
\arguments{
\item{obj}{Object to rotate. Methods are provided for \link{shape}
and \link{transform}.}

\item{around}{Location to rotate around. If missing, the object's
global position is used (see \code{\link[=obj_pos]{obj_pos()}}).}

\item{...}{Passed on to methods.}

\item{radians}{Numeric; rotation angle in radians.}

\item{degrees}{Numeric; rotation angle in degrees. If both are provided,
\code{radians} takes precedence.}

\item{local}{Logical; if \code{TRUE}, rotate in the object's local
coordinate system instead of world coordinates.}
}
\value{
The modified object, invisibly.
}
\description{
\code{obj_rotate()} rotates a \link{shape} or \link{transform}
around a given point. It updates the position and/or angle of the
object, and can optionally recurse into all children of a shape.
}
\examples{
w <- window(bl = 0, tr = pos(10, 10))
rect_obj <- rect(width = 2, height = 2, trans = transform(pos(5, 5)))
w@child(rect_obj)

# Rotate 45 degrees around the center
obj_rotate(rect_obj, around = pos(5, 5), degrees = 45)

}
\seealso{
\code{\link[=obj_scale]{obj_scale()}}, \code{\link[=obj_translate]{obj_translate()}}, \code{\link[=window]{window()}}
}
