% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\name{obj_scale}
\alias{obj_scale}
\title{Scale an object}
\usage{
obj_scale(
  obj,
  around,
  ...,
  size,
  target_size = obj_size(obj) + size,
  scale = target_size/obj_size(obj),
  local = FALSE
)
}
\arguments{
\item{obj}{dispatch arg to scale}

\item{around}{location to scale from}

\item{...}{for extensions}

\item{size, target_size, scale}{These arguments
do much of the same thing. \code{scale} determines the
scaling factor that will be applied to transforms.
target_scale represents the end size value the transform
will have. \code{size} represents a value or delta to change
the objects current size value. Note that for \code{size} and
\code{target_size} are merely reparameteriztions for \code{scale}
which will be the value used by inner functions.}
}
\description{
Scale an object
}
