% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-transform.R
\name{obj_size}
\alias{obj_size}
\title{Query size of an object}
\usage{
obj_size(obj)
}
\arguments{
\item{obj}{Object to query (typically a \code{\link{transform}} or \code{\link{shape}}).}
}
\value{
Numeric size value.
}
\description{
\code{obj_size()} retrieves the size property of a shape or transform.
}
