% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-actions.R
\name{offset}
\alias{offset}
\title{Offset action}
\usage{
offset(by, time, ...)
}
\arguments{
\item{by}{A \code{\link{pos}} delta (or per-time-unit delta); will be scaled
by \code{time@delta} during animation.}

\item{time}{A \code{\link{time}} object controlling duration and easing.}
}
\value{
An \code{\link{action}} object.
}
\description{
\code{offset()} creates an action that moves a \code{\link{shape}} by a continuous
delta per unit time. Over the course of the action's duration,
the total movement is \code{by * time@value}.
}
\seealso{
\code{\link[=translate]{translate()}}, \code{\link{action}}, \code{\link{time}}
}
\keyword{internal}
