% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-pos.R
\name{positions}
\alias{positions}
\title{Combine position vectors}
\usage{
positions(...)
}
\arguments{
\item{...}{One or more \code{\link{pos}} objects or numeric values to combine.}
}
\value{
A \code{\link{pos}} vector.
}
\description{
\code{positions()} combines multiple \code{\link{pos}} objects or vectors
into a single position vector.
}
\examples{
p1 <- pos(1, 2)
p2 <- pos(3, 4)
positions(p1, p2)

}
