% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranim-package.R
\docType{package}
\name{ranim-package}
\alias{ranim}
\alias{ranim-package}
\title{ranim: Hierarchical Shape-Based Animations in R}
\description{
Tools for creating hierarchical, shape-based animations in R. Shapes can contain other shapes and can be manipulated over time with composable "action" objects. Animations are rendered frame-by-frame and can be exported as GIFs.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://example.org/ranim}
  \item Report bugs at \url{https://example.org/ranim/issues}
}

}
\author{
\strong{Maintainer}: Justin Landis \email{jtlandis314@gmail.com}

}
\keyword{internal}
