% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-actions.R
\name{rotate}
\alias{rotate}
\title{Rotation action}
\usage{
rotate(degrees, time, ...)
}
\arguments{
\item{degrees}{Numeric; rotation angle in degrees.}

\item{time}{A \code{\link{time}} object controlling duration and easing.}

\item{radians}{Numeric; rotation angle in radians. If both \code{degrees}
and \code{radians} are provided, \code{radians} takes precedence.}
}
\value{
An \code{\link{action}} object.
}
\description{
\code{rotate()} creates an action that rotates a \code{\link{shape}} around a point
over time. The rotation angle can be specified in degrees or radians.
}
\examples{
w <- window(bl = 0, tr = pos(10, 10))
rect_obj <- rect(width = 2, height = 2, trans = transform(pos(5, 5)))
w@child(rect_obj)
rect_obj@action(rotate(degrees = 360, time = time(2)))

}
\seealso{
\code{\link{action}}, \code{\link{time}}, \code{\link[=translate]{translate()}}, \code{\link[=scale2]{scale2()}}
}
