% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate.R
\name{rotate_local_pos}
\alias{rotate_local_pos}
\title{Rotate a position locally}
\usage{
rotate_local_pos(obj, radians)
}
\arguments{
\item{obj}{A \code{\link{pos}} to rotate.}

\item{radians}{Numeric rotation angle in radians.}
}
\value{
A rotated \code{\link{pos}}.
}
\description{
Rotate a position locally
}
\keyword{internal}
