% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate.R
\name{rotate_pos_around_pos}
\alias{rotate_pos_around_pos}
\title{Rotate a position around another position}
\usage{
rotate_pos_around_pos(obj, around, radians)
}
\arguments{
\item{obj}{A \code{\link{pos}} to rotate.}

\item{around}{A \code{\link{pos}} center point.}

\item{radians}{Numeric rotation angle in radians.}
}
\value{
A rotated \code{\link{pos}}.
}
\description{
Rotate a position around another position
}
\keyword{internal}
