% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate.R
\name{rotate_trans}
\alias{rotate_trans}
\title{Rotate a transform}
\usage{
rotate_trans(trans, around, radians, local = FALSE)
}
\arguments{
\item{trans}{A \code{\link{transform}} object.}

\item{around}{A \code{\link{pos}} point to rotate around.}

\item{radians}{Numeric rotation angle in radians.}

\item{local}{Logical; if \code{TRUE}, rotate in local coordinates.}
}
\value{
The modified \code{transform}, invisibly.
}
\description{
Rotate a transform
}
\keyword{internal}
