% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-actions.R
\name{scale}
\alias{scale}
\title{Scaling action}
\usage{
scale(size, time, ...)
}
\arguments{
\item{size}{Numeric; size change per unit time. Can be positive
(to grow) or negative (to shrink).}

\item{time}{A \code{\link{time}} object controlling duration and easing.}
}
\value{
An \code{\link{action}} object.
}
\description{
\code{scale()} creates an action that continuously scales a \code{\link{shape}}
by a given \code{size} delta per unit time.
}
\details{
This function applies a continuous scaling over time. For scaling
from one specific size to another, use \code{\link[=scale2]{scale2()}} instead.
}
\seealso{
\code{\link[=scale2]{scale2()}}, \code{\link{action}}, \code{\link{time}}, \code{\link[=rotate]{rotate()}}
}
\keyword{internal}
