% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-actions.R
\name{scale2}
\alias{scale2}
\title{Scale action (endpoint specification)}
\usage{
scale2(size, time)
}
\arguments{
\item{size}{Numeric; the amount to add to the current size to reach
the target. (I.e., \code{target_size = current_size + size}.)}

\item{time}{A \code{\link{time}} object controlling duration and easing.}
}
\value{
An \code{\link{action}} object.
}
\description{
\code{scale2()} creates an action that scales a \code{\link{shape}} to a target
size over time. Unlike \code{\link[=scale]{scale()}}, which applies a continuous
delta, \code{scale2()} interpolates from the current size to an absolute
target size.
}
\examples{
w <- window(bl = 0, tr = pos(10, 10))
rect_obj <- rect(width = 1, height = 1, trans = transform(pos(5, 5)))
w@child(rect_obj)

# Scale up by 1 unit over 1 second
rect_obj@action(scale2(size = 1, time = time(1)))

}
\seealso{
\code{\link[=scale]{scale()}}, \code{\link{action}}, \code{\link{time}}, \code{\link[=translate]{translate()}}
}
