% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-actions.R
\name{spawn}
\alias{spawn}
\title{Spawn a clone at action completion}
\usage{
spawn(obj, time)
}
\arguments{
\item{obj}{A \code{\link{shape}} object to clone and spawn.}

\item{time}{A \code{\link{time}} object controlling when the spawn occurs.}
}
\value{
An \code{\link{action}} object.
}
\description{
\code{spawn()} creates an action that clones a shape and adds it as
a child when the action completes.
}
\seealso{
\code{\link[=spawn_from]{spawn_from()}}, \code{\link[=spawn_sibling]{spawn_sibling()}}, \code{\link[=despawn]{despawn()}}
}
\keyword{internal}
