% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-actions.R
\name{spawn_sibling}
\alias{spawn_sibling}
\title{Spawn a sibling shape}
\usage{
spawn_sibling(sib, time, offset = NULL)
}
\arguments{
\item{sib}{A \code{\link{shape}} object to clone and spawn as a sibling.}

\item{time}{A \code{\link{time}} object controlling when the spawn occurs.}

\item{offset}{Optional \code{\link{pos}} giving the relative position of the
sibling.}
}
\value{
An \code{\link{action}} object.
}
\description{
\code{spawn_sibling()} creates an action that clones and adds a shape
as a sibling (to the same parent) when the action completes.
}
\seealso{
\code{\link[=spawn]{spawn()}}, \code{\link[=spawn_from]{spawn_from()}}, \code{\link[=despawn]{despawn()}}
}
\keyword{internal}
