% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render-helpers.R
\name{tl}
\alias{tl}
\alias{tr}
\alias{bl}
\alias{br}
\title{Get corners of a shape}
\usage{
tl(obj)

tr(obj)

bl(obj)

br(obj)
}
\arguments{
\item{obj}{A \code{\link{shape}} object.}
}
\value{
A \code{\link{pos}} object representing the requested corner.
}
\description{
These functions return the corner positions of a shape's bounding box.
}
\details{
\itemize{
\item \code{tl()}: top-left
\item \code{tr()}: top-right
\item \code{bl()}: bottom-left
\item \code{br()}: bottom-right
}
}
\seealso{
\code{\link[=lc]{lc()}}, \code{\link[=rc]{rc()}}, \code{\link[=tc]{tc()}}, \code{\link[=bc]{bc()}}, \code{\link[=center]{center()}}
}
\keyword{internal}
