% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-actions.R
\name{translate}
\alias{translate}
\title{Translation action}
\usage{
translate(to, time)
}
\arguments{
\item{to}{A \code{\link{pos}} giving the target position in world coordinates.}

\item{time}{A \code{\link{time}} object controlling duration and easing.}
}
\value{
An \code{\link{action}} object.
}
\description{
\code{translate()} creates an action that moves a \code{\link{shape}} from its
current position to a target position over time.
}
\examples{
w <- window(bl = 0, tr = pos(10, 10))
rect_obj <- rect(width = 2, height = 2, trans = transform(pos(2, 2)))
w@child(rect_obj)

# Move from (2, 2) to (8, 8) over 2 seconds
rect_obj@action(translate(to = pos(8, 8), time = time(2)))

}
\seealso{
\code{\link[=offset]{offset()}}, \code{\link{action}}, \code{\link{time}}, \code{\link[=rotate]{rotate()}}
}
