% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{translate_trans}
\alias{translate_trans}
\title{Translate a transform}
\usage{
translate_trans(trans, to, local = FALSE)
}
\arguments{
\item{trans}{A \code{\link{transform}} object.}

\item{to}{Target position.}

\item{local}{If \code{FALSE}, update global position; if \code{TRUE}, update local offset.}
}
\value{
The modified transform.
}
\description{
Translate a transform
}
\keyword{internal}
