% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-window.R
\name{window-rotation}
\alias{window-rotation}
\title{Window rotation helper}
\description{
If \code{recursive} is \code{TRUE}, rotate all children; if \code{FALSE}, only rotate
the window itself.
}
\keyword{internal}
